Aspiration windows (#9)
* init

* add asp windows

+++ src/utils/consts.rs
@@ -18,13 +18,13 @@ pub const fn get_piece_type(piece_code: char) -> PieceType {
     }
 }
 
-pub const MAX_DEPTH: i32 = 127;
+pub const MAX_DEPTH: usize = 127;
 
 // Constants which represent evaluation thresholds.
 pub const WORST_EVAL: i32 = -i32::MAX;
 pub const SHALLOWEST_PROVEN_LOSS: i32 = -100000;
-pub const DEEPEST_PROVEN_LOSS: i32 = SHALLOWEST_PROVEN_LOSS + MAX_DEPTH;
-pub const DEEPEST_PROVEN_WIN: i32 = SHALLOWEST_PROVEN_WIN - MAX_DEPTH;
+pub const DEEPEST_PROVEN_LOSS: i32 = SHALLOWEST_PROVEN_LOSS + MAX_DEPTH as i32;
+pub const DEEPEST_PROVEN_WIN: i32 = SHALLOWEST_PROVEN_WIN - MAX_DEPTH as i32;
 pub const SHALLOWEST_PROVEN_WIN: i32 = 100000;
 pub const BEST_EVAL: i32 = i32::MAX;
